Changing usigned int to size_t, protect loop variable

+++ philo/srcs/philo.c
@@ -6,7 +6,7 @@
 /*   By: mbraets <mbraets@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/24 13:44:28 by mbraets           #+#    #+#             */
-/*   Updated: 2022/05/20 17:32:31 by mbraets          ###   ########.fr       */
+/*   Updated: 2022/05/23 12:53:09 by mbraets          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 /**
  * TEST:
- *  5 800 200 -> never die
+ *  5 800 200 200 -> never die
  *  5 800 200 200 7
  *  4 410 200 200 -> never die
  *  4 310 200 100 -> one die
@@ -45,7 +45,7 @@ int	philo_parse(t_data *data, int ac, char **av)
 
 int	philo_init(t_data *data)
 {
-	int	i;
+	size_t	i;
 
 	if (ft_palloc(&data->philos, data->philo_max * sizeof(t_philo *))
 		|| ft_palloc(&data->forks, data->philo_max * sizeof(t_forks)))
@@ -58,7 +58,7 @@ int	philo_init(t_data *data)
 		if (pthread_mutex_init(&data->forks[i], NULL) != 0)
 			return (1);
 		ft_memcpy(data->philos[i], &(t_philo){.index = i, .data = data,
-			.right = &data->forks[i]}, sizeof(t_philo)); // Remove memcpy
+			.right = &data->forks[i]}, sizeof(t_philo));
 		if (pthread_create(&data->philos[i]->thread, NULL, &routine, data->philos[i]) != 0)
 			return (1);
 		if (i == 0)
@@ -72,7 +72,7 @@ int	philo_init(t_data *data)
 
 int	philo_wait_thread(t_data *data)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < data->philo_max)
